+++ src/channel.rs
@@ -1,12 +1,12 @@
 use {marker, private, Channel, Error};
 
 #[doc(hidden)]
-pub trait ChannelSupport<E>: private::Sealed {
-    fn check_channel_is_appropriate(channel: Channel) -> Result<(), Error<E>>;
+pub trait ChannelSupport<CommE, PinE>: private::Sealed {
+    fn check_channel_is_appropriate(channel: Channel) -> Result<(), Error<CommE, PinE>>;
 }
 
-impl<E> ChannelSupport<E> for marker::SingleChannel {
-    fn check_channel_is_appropriate(channel: Channel) -> Result<(), Error<E>> {
+impl<CommE, PinE> ChannelSupport<CommE, PinE> for marker::SingleChannel {
+    fn check_channel_is_appropriate(channel: Channel) -> Result<(), Error<CommE, PinE>> {
         if channel != Channel::Ch0 {
             Err(Error::InvalidChannel)
         } else {
@@ -15,8 +15,8 @@ impl<E> ChannelSupport<E> for marker::SingleChannel {
     }
 }
 
-impl<E> ChannelSupport<E> for marker::DualChannel {
-    fn check_channel_is_appropriate(_channel: Channel) -> Result<(), Error<E>> {
+impl<CommE, PinE> ChannelSupport<CommE, PinE> for marker::DualChannel {
+    fn check_channel_is_appropriate(_channel: Channel) -> Result<(), Error<CommE, PinE>> {
         Ok(())
     }
 }